Countdown Calculator
Finds the difference between dates in Years/Months/Days and also In Days, In hours, In minutes and In seconds. 
It also finds the difference between two different times in Hours/Minutes/Seconds and also In hours, In minutes and In seconds. 
It also finds the date after a given number of days or the a previous date by a number of days.