Sort Generator
The generator sorts an array of words or numbers alphabetically or numerically depending on the choices and inputs given by the user. 
The user gets to choose whether to sort alphabetically or numerically. 
The user also gets to choose what to use to separate the inputs. Whether to use commas or spaces. 
The user also gets to choose whether to sort in descending or ascending order.